requirements.txt
torch
torchvision
grad-cam
streamlit
Pillow
matplotlib
pytorch-grad-cam
matplotlib